+++ ui/src/pages/Dashboard/index.tsx
@@ -1,16 +1,14 @@
 import { Box, Grid, Grow, Typography } from '@material-ui/core'
-import React, { useEffect, useRef, useState } from 'react'
+import React, { useEffect, useRef } from 'react'
 
-import { Event } from 'api/events.type'
 import Paper from 'components-mui/Paper'
 import PaperTop from 'components-mui/PaperTop'
 import { RootState } from 'store'
 import T from 'components/T'
+import Timeline from 'components/Timeline'
 import TotalExperiments from './TotalExperiments'
 import Welcome from './Welcome'
-import api from 'api'
 import genChaosStatePieChart from 'lib/d3/chaosStatePieChart'
-import genEventsChart from 'lib/d3/eventsChart'
 import { getStateofExperiments } from 'slices/experiments'
 import { makeStyles } from '@material-ui/core/styles'
 import { useIntl } from 'react-intl'
@@ -46,33 +44,7 @@ export default function Dashboard() {
   const { stateOfExperiments } = useSelector((state: RootState) => state.experiments)
   const dispatch = useStoreDispatch()
 
-  const chartRef = useRef<HTMLDivElement>(null)
   const chaosStatePieChartRef = useRef<any>(null)
-
-  const [events, setEvents] = useState<Event[] | null>(null)
-
-  const fetchEvents = () => {
-    api.events
-      .events()
-      .then(({ data }) => setEvents(data))
-      .catch(console.error)
-  }
-
-  useEffect(fetchEvents, [])
-
-  useEffect(() => {
-    if (events && events.length) {
-      const chart = chartRef.current!
-
-      genEventsChart({
-        root: chart,
-        events,
-        intl,
-        theme,
-      })
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [events])
 
   useEffect(() => {
     dispatch(getStateofExperiments())
@@ -137,7 +109,7 @@ export default function Dashboard() {
           <Grid item xs={12} md={12} lg={9}>
             <Paper>
               <PaperTop title={T('common.timeline')} />
-              <div ref={chartRef} className={classes.container} />
+              <Timeline className={classes.container} />
             </Paper>
           </Grid>
         </Grid>
